requirements.txt
Flask==0.12.1
transitions==0.5.2
pygraphviz==1.4rc1
python-telegram-bot==5.3.0
lxml
requests
